Move AuthToken module to protocol
