+++ packages/liveblocks-core/src/__tests__/room.test.ts
@@ -1,14 +1,14 @@
 import { rest } from "msw";
 import { setupServer } from "msw/node";
 
-import type { RoomAuthToken } from "../AuthToken";
-import { RoomScope } from "../AuthToken";
 import { LiveList } from "../crdts/LiveList";
 import type { LsonObject } from "../crdts/Lson";
 import { lsonToJson } from "../immutable";
 import * as console from "../lib/fancy-console";
 import type { Json, JsonObject } from "../lib/Json";
 import type { Authentication } from "../protocol/Authentication";
+import type { RoomAuthToken } from "../protocol/AuthToken";
+import { RoomScope } from "../protocol/AuthToken";
 import type { BaseUserMeta } from "../protocol/BaseUserMeta";
 import { ClientMsgCode } from "../protocol/ClientMsg";
 import type { IdTuple, SerializedCrdt } from "../protocol/SerializedCrdt";
