+++ packages/liveblocks-core/src/crdts/__tests__/LiveObject.test.ts
@@ -6,7 +6,7 @@ import {
   prepareStorageTest,
   reconnect,
 } from "../../__tests__/_utils";
-import { RoomScope } from "../../AuthToken";
+import { RoomScope } from "../../protocol/AuthToken";
 import { OpCode } from "../../protocol/Op";
 import type { IdTuple, SerializedCrdt } from "../../protocol/SerializedCrdt";
 import { CrdtType } from "../../protocol/SerializedCrdt";
